[Chore #160806671] Restructure the app directory
