+++ app/api/v2/views_users.py
@@ -2,7 +2,7 @@
 from flask import request
 from flask_restful import Resource
 
-from app.api.v2.model import Users
+from app.api.v2.model_users import Users
 
 class UserRegister(Resource):
     """This class is used to register a new user"""
